refactor(schemas): extract remaining inline schemas for all domains
Add categorySchema.ts, locationSchema.ts, subCategorySchema.ts with domain
schemas. Add replyFeedbackSchema to feedbackSchema.ts. Refactor assetSchema.ts
to import shared schemas from domain files instead of redefining them. Update
category, location, subCategory, and feedback/ReplyModal components to import
from app/schemas/ — no inline z.object() declarations remain in any component.
